upgrade fixes

+++ src/db/migrate/20130226133232_add_environment_default_id_to_content_view.rb
@@ -6,6 +6,8 @@ class AddEnvironmentDefaultIdToContentView < ActiveRecord::Migration
       cv.environment_default_id = env.id
       cv.save!
     end
+    raise "All environments not properly migrated" if KTEnvironment.count != ContentView.where("environment_default_id is not null").count
+
     remove_column :environments, :default_content_view_id
     add_index :content_views, [:environment_default_id]
   end
